issue #120 : add mock for RedisServiceTest
